pyproject.toml
[tool.poetry]
name = "docweave"
version = "0.1.0"
description = "Web application powered by GitHub Copilot CLI that analyzes git repositories and generates beautiful documentation"
authors = ["DocWeave"]
readme = "README.md"
packages = [{ include = "docweave", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
gitpython = "^3.1"
click = "^8.1"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
websockets = "^12.0"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
docweave = "docweave.cli:cli"
